requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
Pillow

category_encoders

matplotlib

numpy

opencv_python

pandas

scikit_image

sklearn
